Port ThreadedWorker to Qt4.
QWaitCondition::wait() without arguments no longer exists, now a QMutex
is required, that must be locked on calling, will be unlocked by the
function, and will be locked again before returning.

Because of this, we cannot continue to use the main loop to obtain the
item, and have to introduce a second loop.