XMLGUI, OH HOW SO WELCOME ARE THOU!
Migrate MainWindow to be a KXmlGuiWindow, not a KMainWindow. We'll see
how it goes when we want to use a toolbar in a non-standard location.

This commit changes the creation of clases to the new KDE4 way, and
drops the actions/slots for configuring the shortcuts, since these are
created automatically by setupGUI(). We also loose the init_menus()
method, since standard actions are added in the right menus, and for our
own we use a minirokui.rc file.