Rewrite the configuration handling of Columns.
Now a single option is used, that contains all the necessary information.
Note that the setup has to move out from __init__, else it does not work
(probably because it's not associated to a view yet). Since KConfig
crashes at the moment with PyKDE, use a fake stub config object.

Also, drop minirokrc now (defaults are in the code now, but they could
still be overriden with a system-wide configuration file AFAIK).