Move the handling of the queue on item re-insertion to RemoveItemsCmd.
This is necessary because the previous state of the queue can only be
meaningfully restored if we are restoring all the removed items in the
same go (otherwise the sorted indexes may not be valid), but insert_items
is used in batches on contiguous rows, so it does not work.

There is one case in which having RemoveItemsCmd remove items from the
queue is not wanted: when doing internal drag and drop, items are removed
with RemoveItemsCmd, but are re-inserted with InsertItemsCmd, which does
not touch the queue at all (and would not have access to the old queue
indexes anyway).

For this, rename the previous RemoveItemsCmd to RemoveItemsCmdNoQueue,
and derive from that a new RemoveItemsCmd that adds the logic to handle
the queue, and make dropMimeData() use RemoveItemsCmdNoQueue.

Finally, all this allows us to get rid of the keep_queue_position_attr
hack in toggle_enqueued_many().