Fix util.RandomOrderedList.extend().
Previously, it would shuffle the sequence received as parameter, and
append it to self, meaning that items being added later would not mix
with already present items.

Now we extend self with the new sequence, and shuffle the result. This
means that the order changes each time we add items, but that should be
okay.