Update Last.fm submission code to play nicely with lastfmsubmitd 1.0.2.
lastfmsubmitd 1.0.2 has reworked the way running a instance of the daemon
as one's user. It no longer reads ~/.lastfmsubmitd.conf, but ~/.<name>/conf,
where name is the name passed to the Client constructor.

We were creating clients with Client('minirok'), hence the configuration was
not found. Use now 'lastfmsubmitd' as the name of the client, so that the
user's default daemon is used.

This commit also updates README.Lastfm, and makes it 1.0.2-centric.