Make the search line in the tree view opt-in.
For users that don't use the search function in the tree view, it can be
very annoying to have Minirok recurse the whole filesystem tree eaach time
it starts.

This commit adds UI to achieve that, by providing an "Enable" button in the
search widget. When it gets enabled, the buttion morphs into a "Stop scan"
button. The latest state is remembered on exit.