tree_view.py: drop empty directories from the model.
This was one of the missing features in this KDirLister-driven version
of the tree view. As it turns out, it's not completely trivial to do,
because whilst the removal is easy, those empty directories may gain
playable content, and KDirLister will report it back to us. However,
those directories don't exist in the model any more, and they have to be
brought back somehow.