tree_view.TreeView: implement auto-expand of items after search.
This was the other piece of missing functionality with respect to
kde4/master. However, we use now a better approach than good-old
e1d5572.

As pointed by David Faure on IRC, invalidateFilter() is not asynchronous,
so it's possible to detect when the filtering has finished. So, we have
the proxy model emit a couple of signals, before setting the new filter,
and after invalidateFilter(), so that the view can close auto-expanded
items, and open new visible items, respectively.