Tree view: fix pressing "Enable scan" button when no directory is loaded.
When no directory is loaded, the button is present but populate_pending
in the tree_view is None, leading to tracebacks. We fix this by making
populate_pending be the empty list instead of None.

Additionally, fix the condition in _set_recurse not to assume that if
self._recurse is already True, there won't be any action to perform if
an "Enable it" message comes. Instead, look at whether there's an
ongoing scan or not.

Finally, swap the order of a couple statements in what seems to me a
slightly better ordering regarding race conditions.