playlist.py: prevent assertion failure on Shift+RightButton.
The code in mousePressEvent() was letting events with modifiers other
than Control fall through to the code path intended for presses with no
modifiers, causing eg. Shift+RightButton to produce an assertion failure
because no items were selected. This commits eats all the presses with a
modifier earlier in the code path, preventing this situation.