scrobble.py: use a Lock + Event pair, rather than a Condition object.
A Condition object does nothing if its notify() method gets called while
there's no thread waiting, which is not what we want: in our case, it's
very likely that slot_engine_status_changed() calls notify(), the thread
gets awaken, sees the need for a submission (but not yet the new playing
track), and while submitting it, slot_new_track() calls notify() again
to get the "now playing" notification done. But since run() was not in
wait() yet, it doesn't get sent (which was the observed behavior).

An Event object, on the other hand, when guarding set() and clear() with
a Lock, provides the semantics we need, because set() can be called
whilst nobody is in wait(), and the next wait() will return immediately.