playlist.py: use full-fledged method definitions for undo() and redo().
In PyQt 4.6, the undo() and redo() methods as defined by InsertItemsCmd
and RemoveItemsCmd (via assigning from the mixin class) do not work. This
is a bug reported at [1] which is fixed already, but we make this change
to ensure we work against the buggy version.

  [1]: http://www.riverbankcomputing.com/pipermail/pyqt/2009-October/024723.html

(Cherry picked from commit f296f818b712ded6dc441d10c5ca5db35d8ecf96)