Stop running _ProcInfoCommonTest with negated _has_psutil.
If psutil is not installed, runnig ProcInfo tests with negated _has_psutil
results in running code that expects psutil, which is not installed.

Renamed ProcessInfoOtherTest to ProcessInfoNoPsutilTest, and made it run
always with _has_psutil = False, which results in duplicated tests if
psutil is not present, but maximizes coverage if it is.