Add tests for the Scrobbler class, plus a bugfix for Scrobbler.configured.
This adds tests for all of the Scrobbler class, including a
large/integration test. Also, tests/__init__.py is updated with code to
have a QApplication available (needed for these tests).

This exposed an unlikely race condition in the code, and I've changed
Scrobbler.configured to be a threading.BoundedSemaphore instead of a
Condition. (The race condition arises, if I recall correctly, if the
preferences dialog is closed at a time when we were not in
configured.wait() yet.)